todoist-tasks/requirements.txt
requests
todoist-python